pyproject.toml
[project]
name = "house-price-pred-ml"
version = "0.1.0"
description = "ML for house-price-pred project"
license = "MIT"
authors = [{ name = "Althaf Muhammad", email = "zoory9900@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "joblib>=1.4.2",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
    "scikit-learn>=1.6.1",
    "uvicorn>=0.34.2",
]

[project.scripts]
house-price-pred-ml = "house_price_pred_ml:main"

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "gdown>=5.2.0",
    "ipykernel>=6.29.5",
    "klib>=1.3.2",
    "matplotlib>=3.10.1",
    "pre-commit>=4.3.0",
    "pytest>=8.3.5",
    "pyyaml>=6.0.2",
    "seaborn>=0.13.2",
    "tabulate>=0.9.0",
    "ty>=0.0.1a14",
]
